coding

+++ nkg_quanly/lib/model/ChartModel.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 class ChartSampleData {
   ChartSampleData({this.x = "", this.y = 0, this.color = Colors.white});
   String x;
-  double y;
+  int y;
   Color color;
 }
 
@@ -13,7 +13,7 @@ class PieCharData {
   PieCharData({this.title = "", this.value = 0, this.color = Colors.white});
 
   String title;
-  double value;
+  int value;
   Color color;
 }
